requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
openai==1.3.5
langchain==0.0.350
langchain-community==0.0.13
python-dotenv==1.0.0
tqdm==4.66.1
nltk==3.8.1
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3